WIP on build
wip

+++ modules/atomic-widgets/prop-types/primitives/number-prop-type.php
@@ -18,6 +18,6 @@ class Number_Prop_Type extends Plain_Prop_Type {
 	}
 
 	protected function sanitize_value( $value ) {
-		return (int) $value;
+		return (float) $value;
 	}
 }
